type-annotated func

+++ 0x00-python_variable_annotations/2-floor.py
@@ -7,7 +7,7 @@ takes a float n as argument and returns the floor of the float"""
 import math
 
 
-def floor(n: float) -> int:
+def floor(n: float) -> float:
     """type-annotated function floor which
     takes a float n as argument and returns the floor of the float"""
     return math.floor(n)
